:bug: fix: re-render on create token solved
